+++ config/routes/api.rb
@@ -168,6 +168,7 @@ namespace :api, defaults: { format: "json" } do
 
     # Non-Resourceful Routes
     get "payments/settings", to: "payment_settings#index"
+    patch "payments/settings/upi", to: "payment_settings#update_upi"
     post "payments/settings/stripe/connect", to: "payment_settings#connect_stripe"
     delete "payments/settings/stripe/disconnect", to: "payment_settings#destroy"
     get "calendars/redirect", to: "calendars#redirect", as: "redirect"