+++ db/migrate/20260428165333_add_settings_to_payments_providers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddSettingsToPaymentsProviders < ActiveRecord::Migration[8.1]
+  def change
+    add_column :payments_providers, :settings, :jsonb, null: false, default: {}
+  end
+end